Attempt to fix match

+++ include/decomp8.h
@@ -5,8 +5,8 @@
 #include "psyq/strings.h"
 
 extern s32 D_80047404;
-extern long D_80047408;
+extern s32 D_80047408;
 
-extern void func_800165CC(u8, int, s32);
+extern void func_800165CC(int, int, s32);
 
 #endif
